Separate application components
