+++ src/components/Home.jsx
@@ -1,84 +1,7 @@
-import React, { useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import PropTypes from 'prop-types';
-import { addBook, removeBook } from '@/redux/books/booksSlice';
+import { useSelector } from 'react-redux';
+import Book from '@/components/Book.jsx';
+import NewBookForm from '@/components/NewBookForm';
 import Styles from '@/styles/Home.module.scss';
-
-const Book = ({
-  bookId, title, author, category,
-}) => {
-  const dispatch = useDispatch();
-
-  const deleteThisBook = () => {
-    dispatch(removeBook(bookId));
-  };
-
-  return (
-    <div className={Styles.book}>
-      <div className={Styles['book-description']}>
-        <div>
-          <div className={Styles.category}>{category}</div>
-          <h2 className={Styles['book-title']}>{title}</h2>
-          <span className={Styles.author}>{author}</span>
-        </div>
-        <ul>
-          <li>
-            <button type="button">Comments</button>
-          </li>
-          <li>
-            <button
-              type="button"
-              onClick={() => { deleteThisBook(); }}
-            >
-              Remove
-            </button>
-          </li>
-          <li>
-            <button type="button">Edit</button>
-          </li>
-        </ul>
-      </div>
-      <div />
-      <div />
-    </div>
-  );
-};
-
-const AddBookForm = () => {
-  const [title, setTitle] = useState('');
-  const [author, setAuthor] = useState('');
-  const dispatch = useDispatch();
-
-  const submitForm = (e) => {
-    e.preventDefault();
-    dispatch(addBook({ title, author }));
-    setTitle('');
-    setAuthor('');
-  };
-
-  return (
-    <div className={Styles['book-form-container']}>
-      <h2>ADD NEW BOOK</h2>
-      <form onSubmit={submitForm}>
-        <input
-          placeholder="Book title"
-          value={title}
-          onChange={(e) => setTitle(e.target.value)}
-          className={Styles['title-input']}
-          required
-        />
-        <input
-          placeholder="Book author"
-          value={author}
-          onChange={(e) => setAuthor(e.target.value)}
-          className={Styles['author-input']}
-          required
-        />
-        <button type="submit">ADD BOOK</button>
-      </form>
-    </div>
-  );
-};
 
 const Home = () => {
   const { books } = useSelector((store) => store.books);
@@ -97,20 +20,9 @@ const Home = () => {
       <div className={Styles.books}>
         {allBooks}
       </div>
-      <AddBookForm />
+      <NewBookForm />
     </>
   );
-};
-
-Book.propTypes = {
-  bookId: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired,
-  author: PropTypes.string.isRequired,
-  category: PropTypes.string,
-};
-
-Book.defaultProps = {
-  category: 'Fiction',
 };
 
 export default Home;
