+++ src/components/NewBookForm.jsx
@@ -1,0 +1,42 @@
+import React, { useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { addBook } from '@/redux/books/booksSlice';
+import Styles from '@/styles/NewBookForm.module.scss';
+
+const NewBookForm = () => {
+  const [title, setTitle] = useState('');
+  const [author, setAuthor] = useState('');
+  const dispatch = useDispatch();
+
+  const submitForm = (e) => {
+    e.preventDefault();
+    dispatch(addBook({ title, author }));
+    setTitle('');
+    setAuthor('');
+  };
+
+  return (
+    <div className={Styles['book-form-container']}>
+      <h2>ADD NEW BOOK</h2>
+      <form onSubmit={submitForm}>
+        <input
+          placeholder="Book title"
+          value={title}
+          onChange={(e) => setTitle(e.target.value)}
+          className={Styles['title-input']}
+          required
+        />
+        <input
+          placeholder="Book author"
+          value={author}
+          onChange={(e) => setAuthor(e.target.value)}
+          className={Styles['author-input']}
+          required
+        />
+        <button type="submit">ADD BOOK</button>
+      </form>
+    </div>
+  );
+};
+
+export default NewBookForm;
