+++ src/styles/Book.module.scss
@@ -1,0 +1,59 @@
+.book {
+  background-color: #fff;
+  width: 100%;
+  max-width: 1000px;
+  padding: 30px;
+}
+
+.category {
+  color: var(--secondary-color);
+  font-size: 15px;
+  font-weight: bold;
+  margin-bottom: 5px;
+}
+
+.book-title {
+  font-size: 25px;
+  font-weight: bold;
+  font-family: 'Roboto Slab', serif;
+}
+
+.author {
+  font-size: 15px;
+  font-weight: lighter;
+  font-family: 'Roboto Slab', serif;
+  color: var(--cool-blue);
+}
+
+.book-description {
+  display: flex;
+  flex-direction: column;
+  gap: 50px;
+
+  ul {
+    list-style: none;
+    display: flex;
+
+    li {
+      border-right: 2px solid var(--secondary-color);
+      padding: 0 10px;
+
+      &:first-child {
+        padding-left: 0;
+      }
+
+      &:last-child {
+        border-right: none;
+      }
+
+      button {
+        font-size: 16px;
+        background: none;
+        border: none;
+        color: var(--cool-blue);
+        cursor: pointer;
+      }
+    }
+  }
+}
+
