+++ src/styles/NewBookForm.module.scss
@@ -1,0 +1,57 @@
+
+.book-form-container {
+  max-width: 1000px;
+  margin: auto;
+  border-top: 1px solid var(--secondary-color);
+  padding: 50px 0;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+
+  h2 {
+    font-size: 25px;
+    color: var(--secondary-color);
+  }
+
+  form {
+    display: flex;
+    gap: 10px;
+
+    * {
+      padding: 10px;
+    }
+
+    .title-input {
+      width: 50%;
+    }
+
+    .author-input {
+      width: 30%;
+    }
+
+    input {
+      border: 1px solid #dadada;
+      outline: none;
+    }
+
+    input::placeholder {
+      color: var(--secondary-color);
+      letter-spacing: 0.5px;
+    }
+
+    button {
+      width: 20%;
+      background-color: var(--primary-color);
+      border: none;
+      color: #fff;
+      border-radius: 5px;
+      cursor: pointer;
+      transition: 0.2s all;
+
+      &:hover {
+        background-color: #fff;
+        color: var(--primary-color);
+      }
+    }
+  }
+}
